requirements.txt
keras==2.10.0
Keras_Preprocessing==1.1.2
matplotlib==3.5.2
numpy==1.23.0
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
Pillow==9.2.0
Pillow==10.3.0
streamlit==1.17.0
tensorflow==2.10.0
